# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "This is the API documentation for the Laravel project using Swagger",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get all todos",
                "operationId": "03dd9189a239d7cc845bdf0628ce4859",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Filter by created date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of todos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Todo"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "operationId": "5d049efb4ec532d10c6860904baef7a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get a specific todo",
                "operationId": "4f2c032d591cfeb608418a9513048d55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update an existing todo",
                "operationId": "35d9b29841b3b3708ad39f31bb5c135b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Todo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a specific todo",
                "operationId": "217d89ecbeb947ce5b5a13b3e4c4775c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todo deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            }
        },
        "/api/todos/{id}/status": {
            "patch": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update the status of a specific todo",
                "operationId": "517cb23e4f28e08899424d4a6f9d64f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the todo to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "New status of the todo",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed",
                                            "canceled"
                                        ],
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "Todo status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "Unable to update Todo status"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Response message",
                                            "type": "string",
                                            "example": "Todo not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "title": "Todo",
                "description": "A Todo object",
                "properties": {
                    "id": {
                        "description": "ID of the todo",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the todo",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the todo",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status of the todo",
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_id": {
                        "description": "ID of the user who created the todo",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}